+++ src/main/java/com/prgrms/ohouse/domain/user/model/UserRepository.java
@@ -1,10 +1,9 @@
 package com.prgrms.ohouse.domain.user.model;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository {
 	Optional<User> findByEmail(String email);
+
+	User save(User user);
 }
